+++ src/torManager/torManager.test.ts
@@ -65,7 +65,7 @@ describe('Tor manager', () => {
         detached: true
       }
     })
-    await torSecondInstance.init()
+    await torSecondInstance.init({})
     await torSecondInstance.kill()
   })
 
@@ -97,5 +97,75 @@ describe('Tor manager', () => {
     console.log(tor.torPassword)
     expect(tor.torHashedPassword).toHaveLength(61)
     expect(tor.torPassword).toHaveLength(32)
+  })
+
+  it('tor spawn repeating 3 times with 1 second timeout and repeating will stop after that', async () => {
+    const torPath = torBinForPlatform()
+    const [controlPort] = await fp(9051)
+    const httpTunnelPort = (await fp(controlPort as number + 1)).shift()
+    const socksPort = (await fp(httpTunnelPort as number + 1)).shift()
+    const libPath = torDirForPlatform()
+    const tor = new Tor({
+      appDataPath: tmpAppDataPath,
+      socksPort,
+      torPath: torPath,
+      controlPort,
+      httpTunnelPort,
+      options: {
+        env: {
+          LD_LIBRARY_PATH: libPath,
+          HOME: tmpAppDataPath
+        },
+        detached: true
+      }
+    })
+
+    await expect(tor.init({ repeat: 3, timeout: 1000 }))
+      .rejects
+      .toThrow('Failed to spawn tor 4 times')
+
+    await tor.kill()
+  })
+
+  it('tor is initializing correctly with 40 seconds timeout', async () => {
+    const torPath = torBinForPlatform()
+    const [controlPort] = await fp(9051)
+    const httpTunnelPort = (await fp(controlPort as number + 1)).shift()
+    const socksPort = (await fp(httpTunnelPort as number + 1)).shift()
+    const libPath = torDirForPlatform()
+    const tor = new Tor({
+      appDataPath: tmpAppDataPath,
+      socksPort,
+      torPath: torPath,
+      controlPort,
+      httpTunnelPort,
+      options: {
+        env: {
+          LD_LIBRARY_PATH: libPath,
+          HOME: tmpAppDataPath
+        },
+        detached: true
+      }
+    })
+
+    await tor.init({ repeat: 3, timeout: 40000 })
+    await tor.kill()
+  })
+
+  it('creates and destroys hidden service', async () => {
+    const tor = await spawnTorProcess(tmpAppDataPath)
+    await tor.init()
+    const hiddenService = await tor.createNewHiddenService(4343, 4343)
+    const status = await tor.destroyHiddenService(hiddenService.onionAddress)
+    expect(status).toBe(true)
+    await tor.kill()
+  })
+
+  it('attempt destroy nonexistent hidden service', async () => {
+    const tor = await spawnTorProcess(tmpAppDataPath)
+    await tor.init()
+    const status = await tor.destroyHiddenService('u2rg2direy34dj77375h2fbhsc2tvxj752h4tlso64mjnlevcv54oaad')
+    expect(status).toBe(false)
+    await tor.kill()
   })
 })