+++ src/nodeTest/nodes.ts
@@ -78,9 +78,9 @@ export class NodeWithoutTor extends LocalNode {
     )
     const communities = new CommunitiesManager(connectonsManager)
     const peerId = await this.getPeer()
-    
+
     // TODO: use this after turning on cert handling again:
-    const bootstrapAddressArrayWs = this.bootstrapMultiaddrs.map((address) => address.replace('wss', 'ws'))
+    // const bootstrapAddressArrayWs = this.bootstrapMultiaddrs.map((address) => address.replace('wss', 'ws'))
     // eslint-disable-next-line
     const certs = {} as CertsData