+++ CHANGELOG.md
@@ -2,12 +2,17 @@
 
 ## [Unreleased]
 
+### Fixed
+
+* Stucking Tor - kill and relaunch the Tor after given timeout
+
 ### Added
 
 * DestroyHiddenService method in torManager
 * CreateNetwork in communitiesManager
 * CreateNetwork api in IOPRoxy
 * Emit general ERROR socket event
+* ConnectionsManager option - allow using tor files (binaries and libs) from waggle
 
 ### Changed