+++ CHANGELOG.md
@@ -6,11 +6,15 @@
 
 * Certificates validation in websocket over tor is temporarly removed - doesn't work with electron
 
+### Fixed
+
+* Stucking Tor - kill and relaunch the Tor after given timeout
 ### Added
 
 * DestroyHiddenService method in torManager
 * CreateNetwork in communitiesManager
 * CreateNetwork api in IOPRoxy
+* ConnectionsManager option - allow using tor files (binaries and libs) from waggle
 
 ## [4.0.1]